Problems
On the ShapeUp document part, I hadn't changed anything but I was getting a problem. ON lines 37 and 43 my computer was saying that it couldn't find the circumfrence method. Therefore I have no idea whether or not any of this is correct so I wasn't able to check my output.